RD-4279: Extracted TerraformModal to a dedicated directory
