+++ widgets/blueprints/src/TerraformModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TerraformModal';
